refactor(ui): split message part rendering
