+++ packages/ui/src/components/message-part/parts/compaction-and-divider.tsx
@@ -1,0 +1,21 @@
+import { useI18n } from "../../../context/i18n"
+import { registerPartComponent } from "../registry"
+
+export function MessageDivider(props: { label: string }) {
+  return (
+    <div data-component="compaction-part">
+      <div data-slot="compaction-part-divider">
+        <span data-slot="compaction-part-line" />
+        <span data-slot="compaction-part-label" class="text-13-regular text-fg-weak">
+          {props.label}
+        </span>
+        <span data-slot="compaction-part-line" />
+      </div>
+    </div>
+  )
+}
+
+registerPartComponent("compaction", function CompactionPartDisplay() {
+  const i18n = useI18n()
+  return <MessageDivider label={i18n.t("ui.messagePart.compaction")} />
+})
