+++ packages/ui/src/components/message-part/parts/index.ts
@@ -1,0 +1,4 @@
+import "./compaction-and-divider"
+import "./reasoning"
+import "./text"
+import "./tool"
