+++ packages/ui/src/components/message-part/tools/index.ts
@@ -1,0 +1,15 @@
+import "./read"
+import "./list"
+import "./glob"
+import "./grep"
+import "./webfetch"
+import "./websearch"
+import "./worktree"
+import "./agent"
+import "./bash"
+import "./edit"
+import "./write"
+import "./apply-patch"
+import "./todowrite"
+import "./question"
+import "./skill"
