+++ packages/ui/src/components/message-part/tools/read.tsx
@@ -1,0 +1,48 @@
+import { createMemo, For } from "solid-js"
+import { getFilename } from "@opencode-ai/core/util/path"
+import { useData } from "../../../context"
+import { useI18n } from "../../../context/i18n"
+import { BasicTool } from "../../basic-tool"
+import { Icon } from "../../icon"
+import { relativizeProjectPath } from "../markdown-render"
+import { ToolRegistry } from "../registry"
+
+ToolRegistry.register({
+  name: "read",
+  render(props) {
+    const data = useData()
+    const i18n = useI18n()
+    const args: string[] = []
+    if (props.input.offset) args.push("offset=" + props.input.offset)
+    if (props.input.limit) args.push("limit=" + props.input.limit)
+    const loaded = createMemo(() => {
+      if (props.status !== "completed") return []
+      const value = props.metadata.loaded
+      if (!value || !Array.isArray(value)) return []
+      return value.filter((p): p is string => typeof p === "string")
+    })
+    return (
+      <>
+        <BasicTool
+          {...props}
+          icon="glasses"
+          trigger={{
+            title: i18n.t("ui.tool.read"),
+            subtitle: props.input.filePath ? getFilename(props.input.filePath) : "",
+            args,
+          }}
+        />
+        <For each={loaded()}>
+          {(filepath) => (
+            <div data-component="tool-loaded-file">
+              <Icon name="enter" />
+              <span>
+                {i18n.t("ui.tool.loaded")} {relativizeProjectPath(filepath, data.directory)}
+              </span>
+            </div>
+          )}
+        </For>
+      </>
+    )
+  },
+})
